+++ app/src/main/java/gpj/androidsearch/NetworkRequest.java
@@ -12,6 +12,7 @@ import org.apache.http.impl.client.DefaultHttpClient;
 import org.apache.http.params.BasicHttpParams;
 import org.apache.http.params.HttpConnectionParams;
 import org.apache.http.params.HttpParams;
+import org.json.JSONException;
 import org.json.JSONObject;
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
@@ -24,23 +25,39 @@ public class NetworkRequest extends AsyncTask<String, Void, JSONObject> {
         return SEARCH_ENDPOINT + URLEncoder.encode(String.valueOf(query), "UTF-8");
     }
 
+    private static String queryFromParams(String... params) {
+        if (params == null || params.length == 0 || params[0] == null) {
+            return "";
+        }
+
+        return params[0];
+    }
+
+    private static JSONObject errorResult(String message) {
+        JSONObject json = new JSONObject();
+        try {
+            json.put("text", message);
+            json.put("image", "");
+        } catch (JSONException e) {
+            Log.e("network_request", "Unable to build error result", e);
+        }
+
+        return json;
+    }
+
     @Override
     protected JSONObject doInBackground(String... params) {
         try {
-            String query;
-            query = params[0];
+            String query = queryFromParams(params);
 
             HttpParams httpParams = new BasicHttpParams();
             HttpConnectionParams.setConnectionTimeout(httpParams,
                     1000);
             HttpConnectionParams.setSoTimeout(httpParams, 1000);
-            //
-            HttpParams p = new BasicHttpParams();
-            // p.setParameter("name", pvo.getName());
-            p.setParameter("user", "1");
+            httpParams.setParameter("user", "1");
 
             // Instantiate an HttpClient
-            HttpClient httpclient = new DefaultHttpClient(p);
+            HttpClient httpclient = new DefaultHttpClient(httpParams);
             String url = buildSearchUrl(query);
             Log.d("url", url);
             HttpGet httpget = new HttpGet(url);
@@ -60,18 +77,21 @@ public class NetworkRequest extends AsyncTask<String, Void, JSONObject> {
 
 
             } catch (ClientProtocolException e) {
-                // TODO Auto-generated catch block
-                e.printStackTrace();
+                Log.e("network_request", "Search protocol error", e);
+                return errorResult("Search request failed");
             } catch (IOException e) {
-                // TODO Auto-generated catch block
-                e.printStackTrace();
+                Log.e("network_request", "Search IO error", e);
+                return errorResult("Search request failed");
+            } catch (JSONException e) {
+                Log.e("network_request", "Search response parse error", e);
+                return errorResult("Search request failed");
             }
 
 
         } catch (Throwable t) {
-
+            Log.e("network_request", "Unexpected search request error", t);
+            return errorResult("Search request failed");
         }
-        return null;
     }
 
     protected void onPostExecute(JSONObject feed) {